test(projection-typeorm): split storeHandles.test.ts into 2 files
hoist shared utils to util.ts
